# Migrate MarshalerUtil to an abstract class