--- a/buildSrc/src/main/kotlin/otel.jmh-conventions.gradle.kts
+++ b/buildSrc/src/main/kotlin/otel.jmh-conventions.gradle.kts
@@ -32,5 +32,7 @@ jmhReport {
 tasks {
   named("jmh") {
     finalizedBy(named("jmhReport"))
+
+    outputs.cacheIf { false }
   }
 }