--- a/exporters/otlp/common/src/main/java/io/opentelemetry/exporter/otlp/internal/MarshalerWithSize.java
+++ b/exporters/otlp/common/src/main/java/io/opentelemetry/exporter/otlp/internal/MarshalerWithSize.java
@@ -13,7 +13,7 @@ abstract class MarshalerWithSize implements Marshaler {
   }
 
   @Override
-  public final int getSerializedSize() {
+  public final int getProtoSerializedSize() {
     return size;
   }
 }