--- a/exporters/otlp/common/src/main/java/io/opentelemetry/exporter/otlp/internal/NoopMarshaler.java
+++ b/exporters/otlp/common/src/main/java/io/opentelemetry/exporter/otlp/internal/NoopMarshaler.java
@@ -9,7 +9,7 @@ enum NoopMarshaler implements Marshaler {
   INSTANCE;
 
   @Override
-  public void writeTo(CodedOutputStream output) {}
+  public void writeTo(Serializer output) {}
 
   @Override
   public int getProtoSerializedSize() {