After all the effort of decoupling proto and export for our main exporters, it pains me that we still depend on the generated protos in the agent due to the use of logging-otlp. So having JSON marshaling would be nice, though not so nice in having to write custom marshalers again for JSON.

Thinking about MarshalerUtil, it seems that it can be made into an abstract class in a way that would allow one Marshaler to be able to write to either binary or json without any significant change in the marshaling. The `Serializer` class here only is essentially `MarshalerUtil` but allowing custom implementations of the actual data writing (in this PR, only added the binary protobuf version). It also has a couple of knobs for serializing the start and end of a message / array, anticipating the JSON versions of those.

Things are a bit more abstract, and there's more virtual method calls. Naturally there is some performance regression. I get ~11.9 baseline and my target was to not slow down more than 12.5. It seems to have landed on 12.1, which is almost in the noise, even more so for the 512-element version. So I think we can be comfortable with the effect on performance.

After
```
Benchmark                                                                (numSpans)  Mode  Cnt       Score       Error   Units
RequestMarshalBenchmarks.marshalCustom                                           16  avgt   10      12.115 ±     0.170   us/op
RequestMarshalBenchmarks.marshalCustom:·gc.alloc.rate                            16  avgt   10    1381.521 ±    18.952  MB/sec
RequestMarshalBenchmarks.marshalCustom:·gc.alloc.rate.norm                       16  avgt   10   26356.128 ±     0.325    B/op
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Eden_Space                   16  avgt   10    1384.085 ±    85.343  MB/sec
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Eden_Space.norm              16  avgt   10   26404.721 ±  1566.848    B/op
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Survivor_Space               16  avgt   10       0.010 ±     0.004  MB/sec
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Survivor_Space.norm          16  avgt   10       0.185 ±     0.074    B/op
RequestMarshalBenchmarks.marshalCustom:·gc.count                                 16  avgt   10     139.000              counts
RequestMarshalBenchmarks.marshalCustom:·gc.time                                  16  avgt   10      83.000                  ms
RequestMarshalBenchmarks.marshalCustom                                          512  avgt   10     418.909 ±     8.223   us/op
RequestMarshalBenchmarks.marshalCustom:·gc.alloc.rate                           512  avgt   10    1221.941 ±    23.681  MB/sec
RequestMarshalBenchmarks.marshalCustom:·gc.alloc.rate.norm                      512  avgt   10  805689.612 ±    11.310    B/op
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Eden_Space                  512  avgt   10    1224.797 ±    71.988  MB/sec
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Eden_Space.norm             512  avgt   10  807541.516 ± 43091.002    B/op
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Survivor_Space              512  avgt   10       0.071 ±     0.095  MB/sec
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Survivor_Space.norm         512  avgt   10      46.642 ±    62.386    B/op
RequestMarshalBenchmarks.marshalCustom:·gc.count                                512  avgt   10     123.000              counts
RequestMarshalBenchmarks.marshalCustom:·gc.time                                 512  avgt   10      82.000                  ms
```

Before
```
Benchmark                                                                (numSpans)  Mode  Cnt       Score       Error   Units
RequestMarshalBenchmarks.marshalCustom                                           16  avgt   10      11.906 ±     0.104   us/op
RequestMarshalBenchmarks.marshalCustom:·gc.alloc.rate                            16  avgt   10    1403.097 ±    12.485  MB/sec
RequestMarshalBenchmarks.marshalCustom:·gc.alloc.rate.norm                       16  avgt   10   26308.168 ±     0.289    B/op
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Eden_Space                   16  avgt   10    1405.722 ±    49.540  MB/sec
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Eden_Space.norm              16  avgt   10   26356.067 ±   785.895    B/op
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Survivor_Space               16  avgt   10       0.034 ±     0.005  MB/sec
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Survivor_Space.norm          16  avgt   10       0.637 ±     0.099    B/op
RequestMarshalBenchmarks.marshalCustom:·gc.count                                 16  avgt   10     141.000              counts
RequestMarshalBenchmarks.marshalCustom:·gc.time                                  16  avgt   10      76.000                  ms
RequestMarshalBenchmarks.marshalCustom                                          512  avgt   10     412.197 ±    23.345   us/op
RequestMarshalBenchmarks.marshalCustom:·gc.alloc.rate                           512  avgt   10    1242.080 ±    64.158  MB/sec
RequestMarshalBenchmarks.marshalCustom:·gc.alloc.rate.norm                      512  avgt   10  805683.147 ±    10.334    B/op
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Eden_Space                  512  avgt   10    1244.567 ±    77.255  MB/sec
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Eden_Space.norm             512  avgt   10  807950.201 ± 57913.258    B/op
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Survivor_Space              512  avgt   10       0.065 ±     0.066  MB/sec
RequestMarshalBenchmarks.marshalCustom:·gc.churn.G1_Survivor_Space.norm         512  avgt   10      41.924 ±    41.964    B/op
RequestMarshalBenchmarks.marshalCustom:·gc.count                                512  avgt   10     102.000              counts
RequestMarshalBenchmarks.marshalCustom:·gc.time                                 512  avgt   10      77.000                  ms
```