@jkwatson 2021-09-03
Do the benchmarks change if you make `Serializer` an interface with default methods, rather than an abstract class? I ask for 2 reasons:

1) concrete inheritance is evil; moving to an interface keeps the code cleaner and allows greater future flexibility
2) I'm honestly curious if the JVM does anything differently between those two approaches.