@jkwatson 2021-09-03
Or, make all the abstract methods into interface methods on a new interface, change the abstract Serializer into a concrete (final) class, and have the new interface impl injected into it to use. Does that change anything with that approach?